requirements.txt
streamlit>=1.30.0
requests>=2.31.0
python-dotenv>=1.0.1
tqdm>=4.66.1
pygments>=2.17.2
